Update to upstream version '6.0.6'
with Debian dir 842ae66b446e68ba24c89c9b98e7e61a6e6a2c68

+++ include/tbl/vrt_stv_var.h
@@ -6,8 +6,8 @@
 
 /*lint -save -e525 -e539 */
 
-VRTSTVVAR(free_space,	BYTES,	long long,	0.)
-VRTSTVVAR(used_space,	BYTES,	long long,	0.)
+VRTSTVVAR(free_space,	BYTES,	int64_t,	0.)
+VRTSTVVAR(used_space,	BYTES,	int64_t,	0.)
 VRTSTVVAR(happy,	BOOL,	unsigned,	0)
 #undef VRTSTVVAR
